Add ignore annotations overlay to all Deployments
This file is a Kustomize overlay that points to the Kustomize base. This overlay includes a patch for the Helm chart base that adds the client.lifecycle.config.k8s.io/mutation: ignore annotation to all Deployment objects. The annotation causes Config Sync to ignore any conflicting changes to this object in the cluster after you have created it.